login dengan session aman gan

+++ delete/config_delete.php
@@ -1,10 +1,9 @@
   
 <?php 
-
+session_start();
 include("./../config/connection.php");
 
-if(isset($_GET['id'])){
-
+if(isset($_GET['id']) && $_SESSION['status'] == "login"){
     $id = $_GET['id'];
     $sql = "DELETE FROM mahasiswa where id=$id";
 
@@ -16,5 +15,5 @@ if(isset($_GET['id'])){
         echo "gagal delete";
     }
 }
-
+    header('Location: ./../read/read-data.php');
 ?>
